+++ extensions/dash-off/src/email-selected-text.tsx
@@ -1,21 +1,35 @@
-import { LaunchType, getPreferenceValues, getSelectedText, launchCommand, showHUD } from "@raycast/api";
+import {
+  LaunchType,
+  Toast,
+  getPreferenceValues,
+  getSelectedText,
+  launchCommand,
+  showHUD,
+  showToast,
+} from "@raycast/api";
 import { sendEmail } from "./lib/gmail";
+import { setGmailToken } from "./lib/withGmailAuth";
 
 export default async function EmailMe() {
   const preferences = getPreferenceValues<Preferences>();
+  await setGmailToken();
   const { defaultAddresses, defaultSubject } = preferences;
 
   try {
     const selectedText = await getSelectedText();
-
+    console.log("selectedText", selectedText);
     try {
+      await showToast({
+        style: Toast.Style.Animated,
+        title: "Sending...",
+        message: "Sending email",
+      });
       const [toAddress, ...BCCAddresses] = defaultAddresses.split(",");
       await sendEmail(defaultSubject, selectedText, toAddress, BCCAddresses);
       showHUD("Email sent!");
     } catch (error) {
-      const err = error as unknown as { response: { body: string } };
-      if (err.response) {
-        return showHUD(`Email failed: ${err.response.body}`);
+      if (error) {
+        return showHUD(`${error}`);
       }
     }
   } catch (error) {