+++ extensions/toggl-track/src/components/CreateTimeEntryForm.tsx
@@ -20,6 +20,7 @@ function CreateTimeEntryForm({ project, description }: { project?: Project; desc
     }
 
     try {
+      await showToast(Toast.Style.Animated, "Starting time entry...");
       await createTimeEntry({
         projectId: selectedProject?.id,
         workspaceId,
@@ -28,10 +29,9 @@ function CreateTimeEntryForm({ project, description }: { project?: Project; desc
         taskId: selectedTask?.id,
         billable,
       });
-      await showToast(Toast.Style.Animated, "Starting time entry...");
-      revalidateRunningTimeEntry();
       await showToast(Toast.Style.Success, "Started time entry");
       navigation.pop();
+      revalidateRunningTimeEntry();
       await clearSearchBar();
     } catch (e) {
       await showToast(Toast.Style.Failure, "Failed to start time entry");