+++ extensions/toggl-track/src/hooks/useTimeEntries.ts
@@ -3,13 +3,13 @@ import { useCachedPromise } from "@raycast/utils";
 import dayjs from "dayjs";
 import { getTimeEntries } from "../api";
 
-export function useTimeEntries(initialExecute = true) {
+export function useTimeEntries() {
   const startDateRef = useRef(dayjs().subtract(1, "week").toDate());
   const endDateRef = useRef(dayjs().toDate());
   const { data, error, isLoading, revalidate } = useCachedPromise(
     () => getTimeEntries({ startDate: startDateRef.current, endDate: endDateRef.current }),
     [],
-    { initialData: [], execute: initialExecute },
+    { initialData: [] },
   );
   return {
     timeEntries: data,