+++ extensions/zed-recent-projects/CHANGELOG.md
@@ -1,7 +1,8 @@
 # zed-recent-projects Changelog
 
-## [Zed Preview Support] - 2024-02-06
+## [Enhancements] - 2024-02-12
 
+- `Search Recent Projects` command now allows to pin and order projects to the top of the list.
 - Added preference for choosing to run commands using Zed or Zed Preview.
 
 ## ["Open With Zed" Command Update] - 2024-01-26